refactor(domain): enum 모듈을 도메인과 통합한다.
