--- core/core-enum/src/main/java/com/giwankim/taxiservice/core/enums/TaxiType.java
@@ -1,7 +1,0 @@
-package com.giwankim.taxiservice.core.enums;
-
-public enum TaxiType {
-  REGULAR,
-  DELUXE,
-  JUMBO
-}
