+++ storage/db-core/build.gradle
@@ -1,6 +1,5 @@
 dependencies {
     implementation project(':core:core-domain')
-    implementation project(':core:core-enum')
 
     implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
     implementation 'org.flywaydb:flyway-core'
